Add methods for "first range check"
Sometimes we may be able to redirect to the resource covered by the first HTTP range that is requested. To
do so, specifically on S3, two conditions must be satisfied:

* The request must be for a single range or for the entire resource - not a multipart (multi-range) request
* The request must be using content from the first interval only - so that we can redirect and have the HTTP client
  retransmit the same HTTP `Range` request to the location of the first interval

This adds methods for implementing both - additionally, if the upstream supports miltirange responses we can
redirect with them as well.